refactor(frontend): one useMedicationResolver — core/fhir survivor, callers repointed (R35)
Two useMedicationResolver implementations with separate module caches
existed (hooks/medication/ and core/fhir/hooks/), causing duplicate
Medication fetches and potentially divergent display names between
SummaryTab and ChartReview.

- Port the hooks/medication capabilities into the core/fhir/hooks
  survivor: contained-resource (#ref) resolution and FHIRResourceContext
  getResource lookup (_include-loaded Medications) before network fetch.
- Repoint all five consumers of the losing implementation:
  EffectivenessMonitoringPanel, MedicationDiscontinuationDialog,
  MedicationEffectivenessDialog, MedicationReconciliationDialog,
  SummaryTab.
- Delete hooks/useMedicationResolver.js (compat shim) and
  hooks/medication/useMedicationResolver.js.
- Fix hooks/medication barrel: its 'default as useMedicationResolver'
  re-export was already broken (module had no default export); now a
  named re-export of the survivor, which hooks/index.js inherits.
- Fix pre-existing no-undef lint errors in two repointed files (missing
  MUI alpha/Divider imports) to keep touched files at 0 eslint errors.
