--- frontend/src/hooks/useMedicationResolver.js
@@ -1,2 +1,0 @@
-// Re-export from new location for backwards compatibility
-export * from './medication/useMedicationResolver';
